--- a/ape_etherscan/client.py
+++ b/ape_etherscan/client.py
@@ -4,15 +4,13 @@ import random
 import time
 from collections.abc import Iterator
 from io import StringIO
-from typing import Optional
+from typing import TYPE_CHECKING, Optional
 
-from ape.api import PluginConfig
 from ape.logging import logger
 from ape.utils import USER_AGENT, ManagerAccessMixin
 from requests import Session
 from yarl import URL
 
-from ape_etherscan.config import EtherscanConfig
 from ape_etherscan.exceptions import (
     ContractNotVerifiedError,
     IncompatibleCompilerSettingsError,
@@ -27,17 +25,22 @@ from ape_etherscan.types import (
 )
 from ape_etherscan.utils import API_KEY_ENV_KEY_MAP
 
+if TYPE_CHECKING:
+    from ape.api import PluginConfig
+
+    from ape_etherscan.config import EtherscanConfig
+
 
 def get_network_config(
-    etherscan_config: EtherscanConfig, ecosystem_name: str, network_name: str
-) -> Optional[PluginConfig]:
+    etherscan_config: "EtherscanConfig", ecosystem_name: str, network_name: str
+) -> Optional["PluginConfig"]:
     if ecosystem_name in etherscan_config:
         return etherscan_config[ecosystem_name].get(network_name)
     return None
 
 
 def get_etherscan_uri(
-    etherscan_config: EtherscanConfig, ecosystem_name: str, network_name: str
+    etherscan_config: "EtherscanConfig", ecosystem_name: str, network_name: str
 ) -> str:
     # Look for explicitly configured Etherscan config
     network_conf = get_network_config(etherscan_config, ecosystem_name, network_name)
@@ -176,7 +179,7 @@ def get_etherscan_uri(
 
 
 def get_etherscan_api_uri(
-    etherscan_config: EtherscanConfig, ecosystem_name: str, network_name: str
+    etherscan_config: "EtherscanConfig", ecosystem_name: str, network_name: str
 ) -> str:
     # Look for explicitly configured Etherscan config
     network_conf = get_network_config(etherscan_config, ecosystem_name, network_name)