### What I did

before:

```
In [1]: %time import ape_etherscan
CPU times: user 987 ms, sys: 140 ms, total: 1.13 s
Wall time: 1.15 s
```

after:

```
In [1]: %time import ape_etherscan
CPU times: user 7.72 ms, sys: 3.29 ms, total: 11 ms
Wall time: 10.7 ms
```

### How I did it

### How to verify it

### Checklist

- [ ] Passes all linting checks (pre-commit and CI jobs)
- [ ] New test cases have been added and are passing
- [ ] Documentation has been updated
- [ ] PR title follows [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) standard (will be automatically included in the changelog)